pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ax-discovery-portal"
version = "0.1.0"
description = "AX BD팀 멀티에이전트 기반 사업기회 포착 엔진"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AX BD Team"}
]
keywords = ["claude", "agent", "mcp", "bd", "discovery"]

dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "atlassian-python-api>=3.41.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "tenacity>=9.0.0",
    "jsonschema>=4.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
ax-portal = "backend.api.main:main"

[project.urls]
Homepage = "https://github.com/AX-BD-Team/ax-discovery-portal"
Repository = "https://github.com/AX-BD-Team/ax-discovery-portal"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
